requirements.txt
fastapi
uvicorn
google-generativeai
python-dotenv
requests
pydantic
python-multipart
pypdf
python-docx
